+++ facebook/login.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/hielfx/gologin/v2"
-	oauth2Login "github.com/hielfx/gologin/v2/oauth2"
+	"github.com/hielfx/gologin"
+	oauth2Login "github.com/hielfx/gologin/oauth2"
 	"golang.org/x/oauth2"
 )