+++ facebook/login_test.go
@@ -7,9 +7,9 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	"github.com/hielfx/gologin/v2"
-	oauth2Login "github.com/hielfx/gologin/v2/oauth2"
-	"github.com/hielfx/gologin/v2/testutils"
+	"github.com/hielfx/gologin"
+	oauth2Login "github.com/hielfx/gologin/oauth2"
+	"github.com/hielfx/gologin/testutils"
 	"github.com/stretchr/testify/assert"
 	"golang.org/x/oauth2"
 )