+++ github/login.go
@@ -6,8 +6,8 @@ import (
 	"net/http"
 	"net/url"
 
-	"github.com/hielfx/gologin/v2"
-	oauth2Login "github.com/hielfx/gologin/v2/oauth2"
+	"github.com/hielfx/gologin"
+	oauth2Login "github.com/hielfx/gologin/oauth2"
 	"github.com/google/go-github/github"
 	"golang.org/x/oauth2"
 )