+++ oauth2/login_test.go
@@ -8,8 +8,8 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/hielfx/gologin/v2"
-	"github.com/hielfx/gologin/v2/testutils"
+	"github.com/hielfx/gologin"
+	"github.com/hielfx/gologin/testutils"
 	"github.com/stretchr/testify/assert"
 	"golang.org/x/oauth2"
 )