+++ tumblr/login.go
@@ -4,8 +4,8 @@ import (
 	"errors"
 	"net/http"
 
-	"github.com/hielfx/gologin/v2"
-	oauth1Login "github.com/hielfx/gologin/v2/oauth1"
+	"github.com/hielfx/gologin"
+	oauth1Login "github.com/hielfx/gologin/oauth1"
 	"github.com/dghubble/oauth1"
 )