+++ twitter/token_test.go
@@ -9,9 +9,9 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/hielfx/gologin/v2"
-	oauth1Login "github.com/hielfx/gologin/v2/oauth1"
-	"github.com/hielfx/gologin/v2/testutils"
+	"github.com/hielfx/gologin"
+	oauth1Login "github.com/hielfx/gologin/oauth1"
+	"github.com/hielfx/gologin/testutils"
 	"github.com/dghubble/oauth1"
 	"github.com/stretchr/testify/assert"
 )